Pad Layout and Information Folder Created
Made an information folder and finally made a working layout for all
pads in the "Layout Spreadsheet" in the information folder. All samples
will be rearranged shortly. The PDF is a copy of the spreadsheet and
looks kind of rough, but it's usable to get the information across. Ask
me if you have any questions about it.