All pads in Outline completed, sample organization to come
I finally finished all of the pads, so now I'm going to go back in
Ableton and correct those samples so that they work correctly. Also, I
will try to change order of samples to match the pdf. Trigger types will
need to be changed when I do this as well. Cheers.