New Samples Put On Pads
I subbed the old ones for the new ones, and now everything is peachy
keen. Next on the to-do list is set default trigger types in the json
file for each sample and to organize pads according to the pdf layout.
Also...We should do the default screen to have all blank samples and
then you can pick the Pop Culture folder to set pads. Cheers.